index/constraint management for embedded, new exclusion filter
