+++ lib/neo4j-server.rb
@@ -5,6 +5,7 @@ require 'faraday_middleware/multi_json'
 require 'neo4j-server/resource'
 require 'neo4j-server/cypher_node'
 require 'neo4j-server/cypher_label'
+require 'neo4j-server/label'
 require 'neo4j-server/cypher_session'
 require 'neo4j-server/cypher_node_uncommited'
 require 'neo4j-server/cypher_relationship'
